+++ src/main/java/kitchenpos/menus/tobe/domain/menu/Menu.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menu;
 
 import jakarta.persistence.*;
 import kitchenpos.exception.CanNotChangeDisplay;
