+++ src/main/java/kitchenpos/menus/tobe/domain/menu/MenuNameFactory.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menu;
 
 import jakarta.transaction.Transactional;
 import kitchenpos.exception.IllegalNameException;
