+++ src/main/java/kitchenpos/menus/tobe/domain/menugroup/JpaMenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menugroup;
 
 import org.springframework.data.jpa.repository.JpaRepository;
