+++ src/main/java/kitchenpos/menus/tobe/domain/menugroup/MenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menugroup;
 
 import java.util.List;
 import java.util.Optional;
