+++ src/test/java/kitchenpos/menu/tobe/domain/InMemoryMenuRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.menu.tobe.domain;
 
-import kitchenpos.menus.tobe.domain.Menu;
-import kitchenpos.menus.tobe.domain.MenuRepository;
+import kitchenpos.menus.tobe.domain.menu.Menu;
+import kitchenpos.menus.tobe.domain.menu.MenuRepository;
 
 import java.util.*;
